worker/requirements.txt
python-dotenv==1.0.0
loguru==0.7.2
psycopg2==2.9.9
alive-progress==3.1.4